refactor(frontend): remove deprecated types barrel
Delete frontend/src/types.ts — a two-line @deprecated re-export shim
pointing to ./types/index.js. Migrate all 36 import sites to import
directly from the appropriate domain file:

- ../types/runtime.js  — RecentEvent, RateLimits, WorkflowColumn,
                         RuntimeIssueView, WebhookHealth, RuntimeSnapshot,
                         StallEventView, SystemHealth
- ../types/issues.js   — AttemptSummary, SteerIssueResponse, IssueDetail,
                         AbortIssueResponse, AttemptCheckpointRecord,
                         AttemptRecord, REASONING_EFFORT_OPTIONS
- ../types/setup.js    — LinearProject, GitContextResponse, SetupStatus,
                         ActiveBranchView, GitCommitView, GitPullView,
                         GitRepoView
- ../types/config.js   — TrackedPrRecord, AuditRecord, PromptTemplate,
                         RuntimeInfo
- ../types/codex.js    — all Codex* types
- ../types/notifications.js — all Notification* types
- ../types/observability.js — ObservabilitySummary
- ../types/workspace.js    — WorkspaceInventoryResponse,
                             WorkspaceInventoryEntry

The barrel had no purpose beyond indirection; types/index.ts already
aggregates all domain files for consumers that need the full set.
