+++ frontend/src/api.ts
@@ -3,7 +3,10 @@ import type {
   AttemptCheckpointRecord,
   AttemptRecord,
   AttemptSummary,
-  AuditRecord,
+  IssueDetail,
+  SteerIssueResponse,
+} from "./types/issues.js";
+import type {
   CodexAdminSnapshotResponse,
   CodexCapabilities,
   CodexAccountLoginStartResponse,
@@ -17,21 +20,17 @@ import type {
   CodexThreadReadResponse,
   CodexThreadListResponse,
   CodexUserInputRequestListResponse,
-  GitContextResponse,
-  IssueDetail,
-  LinearProject,
+} from "./types/codex.js";
+import type { GitContextResponse, LinearProject, SetupStatus } from "./types/setup.js";
+import type { AuditRecord, PromptTemplate, RuntimeInfo, TrackedPrRecord } from "./types/config.js";
+import type { RuntimeSnapshot } from "./types/runtime.js";
+import type {
   NotificationReadResponse,
   NotificationsListResponse,
   NotificationsReadAllResponse,
-  ObservabilitySummary,
-  PromptTemplate,
-  RuntimeInfo,
-  RuntimeSnapshot,
-  SetupStatus,
-  SteerIssueResponse,
-  TrackedPrRecord,
-  WorkspaceInventoryResponse,
-} from "./types";
+} from "./types/notifications.js";
+import type { ObservabilitySummary } from "./types/observability.js";
+import type { WorkspaceInventoryResponse } from "./types/workspace.js";
 import { getReadAccessToken, getWriteAccessToken } from "./access-token";
 
 const BASE = "";
