+++ frontend/src/components/system-health-badge.ts
@@ -1,4 +1,4 @@
-import type { SystemHealth } from "../types.js";
+import type { SystemHealth } from "../types/runtime.js";
 import { formatRelativeTime } from "../utils/format.js";
 
 /**
