--- frontend/src/types.ts
@@ -1,2 +1,0 @@
-/** @deprecated Import from `./types/<domain>.js` instead. */
-export * from "./types/index.js";
